+++ server/src/models/analytic_unit_cache_model.ts
@@ -1,8 +1,6 @@
 import { AnalyticUnitId, AnalyticUnit } from './analytic_units';
 import { Collection, makeDBQ } from '../services/data_service';
 
-import { SerializedCache } from './types';
-
 import * as _ from 'lodash';
 
 
@@ -12,6 +10,61 @@ const MILLISECONDS_IN_INDEX = 60000;
 
 type FindManyQuery = {
   _id: { $in: AnalyticUnitId[] }
+};
+
+export type SerializedCache = {
+  id: string,
+  data?: AnomalyData | PatternData | ThresholdData
+};
+
+export type AnomalyData = {
+  alpha: number,
+  confidence: number,
+  enableBounds: string,
+  timeStep: number,
+  segments: AnomalySegment[]
+};
+
+export type ThresholdData = {
+  value: number,
+  condition: string,
+  timeStep: number
+};
+
+export type PatternData = GeneralPatternData & Partial<ModelData>;
+
+export type AnomalySegment = {
+  from: number,
+  to: number,
+  data: number[]
+};
+
+export type GeneralPatternData = {
+  timeStep: number,
+  patternCenter: number[],
+  patternModel: number[],
+  convolveMax: number,
+  convolveMin: number,
+  windowSize: number,
+  convDelMin: number,
+  convDelMax: number
+};
+
+export type ModelData = { confidence: number } & (TriangleData | JumpData | DropData);
+
+export type TriangleData = {
+  heightMax: number,
+  heightMin: number
+};
+
+export type JumpData = {
+  jumpHeight: number,
+  jumpLength: number
+};
+
+export type DropData = {
+  dropHeight: number,
+  dropLength: number
 };
 
 export class AnalyticUnitCache {
